Use GraphQL subscriptions for the game
Thee is a lot of code here, but essentially I mostly added an ORM to manage the database and created some base resolvers.

This is all untested, I will be putting a bit more time on it in the coming days as that is all the time I have for today.

Check the comments on the code for more information on the code.